+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 - Added support for bidirectional translations
+  - Introduced `-r`/`--reverse` option to influence direction
 - Use `getopts` crate for argument parsing